+++ discoverer/tarantool.go
@@ -7,12 +7,22 @@ import (
 
 	"github.com/tarantool/go-discovery"
 	"github.com/tarantool/go-tarantool/v2"
-	"github.com/tarantool/tt/lib/cluster"
+
+	"github.com/tarantool/go-storage"
+	tcsstorage "github.com/tarantool/go-storage/driver/tcs"
 )
+
+// TarantoolClient is the interface required to create a TCS-backed storage.
+type TarantoolClient interface {
+	tarantool.Doer
+	NewWatcher(key string, callback tarantool.WatchCallback) (tarantool.Watcher, error)
+}
 
 // Tarantool discovers a list of instance configurations from TcS.
 type Tarantool struct {
-	conn   tarantool.Doer
+	// conn obtains information from TcS.
+	conn TarantoolClient
+	// prefix is the configuration prefix for TcS.
 	prefix string
 }
 
@@ -20,7 +30,7 @@ type Tarantool struct {
 var ErrMissingTarantool = errors.New("no Tarantool connector was applied")
 
 // NewTarantool create decorator with Discovery method.
-func NewTarantool(conn tarantool.Doer, prefix string) *Tarantool {
+func NewTarantool(conn TarantoolClient, prefix string) *Tarantool {
 	return &Tarantool{
 		conn:   conn,
 		prefix: prefix,
@@ -33,28 +43,33 @@ func (t *Tarantool) Discovery(ctx context.Context) ([]discovery.Instance, error)
 		return nil, ErrMissingTarantool
 	}
 
+	st := storage.NewStorage(tcsstorage.New(t.conn))
+
 	for {
 		timeout, err := checkTimeout(ctx)
 		if err != nil {
 			return nil, err
 		}
+		select {
+		case <-ctx.Done():
+			return nil, ctx.Err()
+		default:
+		}
 
-		dataCollector := cluster.NewTarantoolAllCollector(t.conn, t.prefix, timeout)
-		collector := cluster.NewYamlCollectorDecorator(dataCollector)
-		config, err := collector.Collect()
+		subCtx, cancel := context.WithTimeout(ctx, timeout)
+		instances, err := buildInstances(subCtx, st, t.prefix)
+		cancel()
+
 		if err != nil {
 			if errors.Is(err, context.DeadlineExceeded) {
-				// Too small timeout? We could retry until the main ctx is not expired.
+				// Too small timeout? We could retry until the main ctx is not
+				// expired.
 				continue
-			}
-			var noConfig cluster.CollectEmptyError
-			if errors.As(err, &noConfig) {
-				return nil, nil
 			}
 
 			return nil, fmt.Errorf("failed to get data from tarantool: %w", err)
 		}
 
-		return parseConfig(config)
+		return instances, nil
 	}
 }